\name{Delaporte-deprecated}
\alias{Delaporte-deprecated}
\title{Deprecated functions in package \sQuote{Delaporte}}

\description{
  These functions are provided for compatibility with older versions
  of \sQuote{Delaporte} only, and may be defunct at the next release.
}

\details{
  
  The following functions are deprecated and will be made defunct; use
  the replacement indicated below:
    \itemize{
      
      \item{In the \code{\link{qdelap}} and \code{\link{rdelap}} functions, the \code{old = TRUE} method is now deprecated. Use \code{old = FALSE} in situations where the approximation is preferred to the exact computation. Future releases may remove the functionality.}
      
    }
}