% Copyright (c) 2013, Avraham Adler All rights reserved
% SPDX-License-Identifier: BSD-2-Clause

\name{Delaporte-internal}

% These are the internal C functions called by the package
\alias{ddelap_C}
\alias{pdelap_C}
\alias{qdelap_C}
\alias{rdelap_C}
\alias{MoMdelap_C}
\alias{gOMPT_C}

\title{Internal Delaporte Functions}
\description{
  Internal Delaporte functions
}
\details{
  These are not to be called directly by the user.
}

\keyword{internal}
