\name{Delaporte-defunct}
\alias{Delaporte-defunct}
\title{Defunct functions in package \sQuote{Delaporte}}

\description{These functions are defunct and no longer available.}

\details{The following functions are defunct; use the replacement indicated below:
    \itemize{
      \item{As of version 5.0.0, the \code{old = TRUE} method for the \code{\link{qdelap}} and \code{\link{rdelap}} functions is defunct and has been removed.}
    }
}