\name{Delaporte-package}
\alias{Delaporte-package}
\alias{Delaporte}
\docType{package}
\title{
\packageTitle{Delaporte}
}
\description{
\packageDescription{Delaporte}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{Delaporte}
\packageIndices{Delaporte}
}
\author{
\packageAuthor{Delaporte}

Maintainer: \packageMaintainer{Delaporte}
}
\keyword{package}
\keyword{distribution}
