% Copyright (c) 2013, Avraham Adler All rights reserved
% SPDX-License-Identifier: BSD-2-Clause

\name{Delaporte-package}
\alias{Delaporte-package}
\docType{package}
\title{
\packageTitle{Delaporte}
}
\description{
\packageDescription{Delaporte}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{Delaporte}
\packageIndices{Delaporte}
}
\author{
\packageAuthor{Delaporte}

Maintainer: \packageMaintainer{Delaporte}
}

\keyword{package}
\keyword{distribution}
