\name{Delaporte-package}
\alias{Delaporte-package}
\docType{package}
\title{\packageTitle{Delaporte}}
\description{\packageDescription{Delaporte}}
\details{
\tabular{ll}{
Package: \tab Delaporte\cr
Type: \tab Package\cr
Version: \tab 6.0.0\cr
Date: \tab 2016-03-31\cr
License: \tab BSD_2_clause \cr
}
}
\author{Maintainer: Avraham Adler \email{Avraham.Adler@gmail.com}}
\keyword{package}
\keyword{distribution}