\name{Delaporte-package}
\alias{Delaporte-package}
\docType{package}
\title{
Statistical functions for the Delaporte distribution
}
\description{
Provides probability mass, distribution, and quantile functions, as well as a random number generator, for the Delaporte distribution.
}
\details{
\tabular{ll}{
Package: \tab Delaporte\cr
Type: \tab Package\cr
Version: \tab 1.0-1\cr
Date: \tab 2013-09-01\cr
License: \tab (>= 2) | LGPL (>= 3) \cr
}
This package provides the standard probability mass, distribution, and quantile functions, as well as a random number generator, for the Delaporte distribution. It is used similarly to the distributions in \code{stats}.
}
\author{
Avraham Adler

Maintainer: Avraham Adler <Avraham.Adler@gmail.com>
}
\keyword{package}
\keyword{distribution}
