\name{Delaporte-package}
\alias{Delaporte-package}
\docType{package}
\title{
Statistical functions for the Delaporte distribution
}
\description{
Provides probability mass, distribution, quantile, random number generation, and method of moments estimation functions for the Delaporte distribution.
}
\details{
\tabular{ll}{
Package: \tab Delaporte\cr
Type: \tab Package\cr
Version: \tab 2.1-0\cr
Date: \tab 2014-04-06\cr
License: \tab GPL (>= 2) | LGPL (>= 3) \cr
}
Provides probability mass, distribution, quantile, random number generation, and method of moments estimation functions for the Delaporte distribution. It is used similarly to the distributions in \code{stats}.
}
\author{
Avraham Adler

Maintainer: Avraham Adler <Avraham.Adler@gmail.com>
}
\keyword{package}
\keyword{distribution}